# docs/swagger.yaml
basePath: /
definitions:
  apierrors.ErrorBody:
    properties:
      code:
        allOf:
        - $ref: '#/definitions/apierrors.ErrorCode'
        example: validation_error
      message:
        example: Invalid email format
        type: string
    type: object
  apierrors.ErrorCode:
    enum:
    - validation_error
    - email_already_exists
    - invalid_credentials
    - email_not_verified
    - unauthorized
    - invalid_refresh_token
    - session_expired
    - user_blocked
    - invalid_verification_token
    - verification_token_expired
    - internal_error
    - not_found
    type: string
    x-enum-varnames:
    - CodeValidationError
    - CodeEmailAlreadyExists
    - CodeInvalidCredentials
    - CodeEmailNotVerified
    - CodeUnauthorized
    - CodeInvalidRefreshToken
    - CodeSessionExpired
    - CodeUserBlocked
    - CodeInvalidVerificationToken
    - CodeVerificationTokenExpired
    - CodeInternalError
    - CodeNotFound
  apierrors.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/apierrors.ErrorBody'
    type: object
  auth.AuthResponse:
    properties:
      tokens:
        $ref: '#/definitions/auth.TokensResponse'
      user:
        $ref: '#/definitions/models.UserResponse'
    type: object
  auth.LoginRequest:
    properties:
      email:
        example: alex@gmail.com
        type: string
      password:
        example: "1234"
        type: string
    type: object
  auth.LogoutRequest:
    properties:
      refresh_token:
        example: dGhpcyBpcyBhIHJlZnJlc2ggdG9rZW4...
        type: string
    type: object
  auth.MeResponse:
    properties:
      user:
        $ref: '#/definitions/models.UserResponse'
    type: object
  auth.MessageResponse:
    properties:
      message:
        type: string
    type: object
  auth.RefreshRequest:
    properties:
      refresh_token:
        example: dGhpcyBpcyBhIHJlZnJlc2ggdG9rZW4...
        type: string
    type: object
  auth.RegisterRequest:
    properties:
      display_name:
        example: John Doe
        type: string
      email:
        example: user@example.com
        type: string
      password:
        example: strong_password123
        type: string
    type: object
  auth.RegisterResponse:
    properties:
      message:
        type: string
    type: object
  auth.ResendVerificationRequest:
    properties:
      email:
        type: string
    type: object
  auth.TokensResponse:
    properties:
      access_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      expires_in:
        example: 900
        type: integer
      refresh_token:
        example: dGhpcyBpcyBhIHJlZnJlc2ggdG9rZW4...
        type: string
    type: object
  auth.VerifyEmailRequest:
    properties:
      token:
        type: string
    type: object
  models.UserResponse:
    properties:
      display_name:
        example: John Doe
        type: string
      email:
        example: user@example.com
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      role:
        example: user
        type: string
      status:
        example: active
        type: string
    type: object
  orders.listResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/orders.orderResponse'
        type: array
      pagination:
        $ref: '#/definitions/orders.paginationResponse'
    type: object
  orders.orderResponse:
    properties:
      candle_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      main_position:
        type: boolean
      order_type:
        type: string
      price:
        type: number
      quantity:
        type: number
      side:
        type: string
      symbol:
        type: string
      timestamp:
        type: integer
      timestamp_str:
        type: string
    type: object
  orders.paginationResponse:
    properties:
      has_more:
        type: boolean
      limit:
        type: integer
      next_cursor:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@alumieye.com
    name: API Support
  description: ALumiEye Backend Authentication API for the MVP
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: ALumiEye API
  version: "1.0"
paths:
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate a user with email and password
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            $ref: '#/definitions/auth.AuthResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/apierrors.ErrorResponse'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/apierrors.ErrorResponse'
        "403":
          description: Account blocked
          schema:
            $ref: '#/definitions/apierrors.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apierrors.ErrorResponse'
      summary: Login user
      tags:
      - auth
  /api/v1/auth/logout:
    post:
      consumes:
      - application/json
      description: Revoke the current session/refresh token
      parameters:
      - description: Refresh token to revoke
        in: body
        name: request
        schema:
          $ref: '#/definitions/auth.LogoutRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Logout successful
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apierrors.ErrorResponse'
      summary: Logout user
      tags:
      - auth
  /api/v1/auth/refresh:
    post:
      consumes:
      - application/json
      description: Exchange a valid refresh token for a new access token
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Tokens refreshed successfully
          schema:
            $ref: '#/definitions/auth.AuthResponse'
        "401":
          description: Invalid or expired refresh token
          schema:
            $ref: '#/definitions/apierrors.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apierrors.ErrorResponse'
      summary: Refresh access token
      tags:
      - auth
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user account with email and password. User must verify
        email before logging in.
      parameters:
      - description: Registration details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Registration successful
          schema:
            $ref: '#/definitions/auth.RegisterResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/apierrors.ErrorResponse'
        "409":
          description: Email already exists
          schema:
            $ref: '#/definitions/apierrors.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apierrors.ErrorResponse'
      summary: Register a new user
      tags:
      - auth
  /api/v1/auth/resend-verification-email:
    post:
      consumes:
      - application/json
      description: Send a new verification email if the account exists and is not
        verified
      parameters:
      - description: Email address
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.ResendVerificationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Generic success message
          schema:
            $ref: '#/definitions/auth.MessageResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/apierrors.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apierrors.ErrorResponse'
      summary: Resend verification email
      tags:
      - auth
  /api/v1/auth/verify-email:
    post:
      consumes:
      - application/json
      description: Verify email address using the token sent by email
      parameters:
      - description: Verification token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.VerifyEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Email verified
          schema:
            $ref: '#/definitions/auth.MessageResponse'
        "400":
          description: Invalid token
          schema:
            $ref: '#/definitions/apierrors.ErrorResponse'
        "410":
          description: Token expired
          schema:
            $ref: '#/definitions/apierrors.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apierrors.ErrorResponse'
      summary: Verify email
      tags:
      - auth
  /api/v1/me:
    get:
      description: Get the profile of the currently authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: User profile
          schema:
            $ref: '#/definitions/auth.MeResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apierrors.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apierrors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - user
  /api/v1/orders:
    get:
      description: List trade orders from eyebroker with optional filters and cursor
        pagination
      parameters:
      - description: Filter by asset symbol (e.g. ETHUSDT)
        in: query
        name: symbol
        type: string
      - description: 'Filter by side: buy or sell'
        in: query
        name: side
        type: string
      - description: Filter by order type (e.g. market, limit)
        in: query
        name: order_type
        type: string
      - description: Start of time range GMT+7 (YYYY-MM-DD HH:MM:SS)
        in: query
        name: from
        type: string
      - description: End of time range GMT+7 (YYYY-MM-DD HH:MM:SS)
        in: query
        name: to
        type: string
      - description: Page size (default 20, max 100)
        in: query
        name: limit
        type: integer
      - description: Cursor from previous response for next page
        in: query
        name: cursor
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/orders.listResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apierrors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apierrors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apierrors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List orders
      tags:
      - orders
  /api/v1/users/{id}:
    get:
      description: Get a user's public information by their ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User info
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/apierrors.ErrorResponse'
      summary: Get user by ID
      tags:
      - users
  /health:
    get:
      description: Check if the API is running
      produces:
      - application/json
      responses:
        "200":
          description: API is healthy
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Health check
      tags:
      - system
securityDefinitions:
  BearerAuth:
    description: 'Type: Bearer {your_token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"